Add generators for epsg checks
