+++ threedi_modelchecker/config.py
@@ -40,6 +40,8 @@ from .checks.cross_section_definitions import (
 from .checks.factories import (
     ForeignKeyCheckSetting,
     generate_enum_checks,
+    generate_epsg_geom_checks,
+    generate_epsg_raster_checks,
     generate_foreign_key_checks,
     generate_geometry_checks,
     generate_geometry_type_checks,
@@ -85,7 +87,7 @@ from .checks.other import (
     UsedSettingsPresentCheck,
     UsedSettingsPresentCheckSingleTable,
 )
-from .checks.raster import (  # RasterHasMatchingEPSGCheck,
+from .checks.raster import (
     GDALAvailableCheck,
     RasterCompressionUsedCheck,
     RasterExistsCheck,
@@ -2462,11 +2464,6 @@ CHECKS += [
         min_value=-9998.0,
         max_value=8848.0,
     ),
-    # RasterHasMatchingEPSGCheck(
-    #     error_code=797,
-    #     level=CheckLevel.WARNING,
-    #     column=models.ModelSettings.dem_file,
-    # ),
     RasterGridSizeCheck(
         error_code=798,
         column=models.ModelSettings.dem_file,
@@ -3474,6 +3471,29 @@ level_map_fk_check = {
 
 unique_columns = [models.BoundaryCondition1D.connection_node_id]
 
+raster_columns = [
+    models.ModelSettings.dem_file,
+    models.GroundWater.equilibrium_infiltration_rate_file,
+    models.ModelSettings.friction_coefficient_file,
+    models.InitialConditions.initial_groundwater_level_file,
+    models.InitialConditions.initial_water_level_file,
+    models.GroundWater.groundwater_hydraulic_conductivity_file,
+    models.GroundWater.groundwater_impervious_layer_level_file,
+    models.GroundWater.infiltration_decay_period_file,
+    models.GroundWater.initial_infiltration_rate_file,
+    models.GroundWater.leakage_file,
+    models.GroundWater.phreatic_storage_capacity_file,
+    models.Interflow.hydraulic_conductivity_file,
+    models.Interflow.porosity_file,
+    models.SimpleInfiltration.infiltration_rate_file,
+    models.SimpleInfiltration.max_infiltration_volume_file,
+    models.Interception.interception_file,
+    models.VegetationDrag.vegetation_height_file,
+    models.VegetationDrag.vegetation_stem_count_file,
+    models.VegetationDrag.vegetation_stem_diameter_file,
+    models.VegetationDrag.vegetation_drag_coefficient_file,
+]
+
 
 class Config:
     """Collection of checks
@@ -3536,7 +3556,10 @@ class Config:
                     message=f"{model.id.name} must be a positive signed 32-bit integer.",
                 )
             ]
-            # self.checks += generate_epsg_geom_checks()
+            self.checks += generate_epsg_geom_checks(model.__table__, error_code=9)
+            self.checks += generate_epsg_raster_checks(
+                model.__table__, raster_columns, error_code=10
+            )
 
         self.checks += CHECKS
         if not self.allow_beta_features:
